feat(plan133 Ф.1): nova_int=intptr_t, nova_uint=uintptr_t; remove usize/isize
- nova_int: int64_t -> intptr_t (signed address-sized, Go C-era intgo precedent)
- nova_uint: uint64_t -> uintptr_t (unsigned address-sized) — new typedef added
- separate int/i64 and uint/u64 C-type mappings in emit_c.rs + external_registry.rs
- usize/isize: removed from Nova type system — error with hint 'use int (Plan 133)'
- removed from BUILTIN_TYPE_NAMES, TyCat::Int, is_primitive_leaf, auto_derive, verify

+++ compiler-codegen/nova_rt/cast.h
@@ -167,12 +167,11 @@ static inline int64_t nova_int_from_f64_bits(double f) {
     return (int64_t)u;
 }
 
-/* Plan 70.5 Q2: int → uint saturation (negative → 0, D54 precedent).
- * Uses int64_t directly (nova_int = typedef int64_t, defined later in nova_rt.h).
- * i64 max < u64 max — no upper saturation needed.
+/* Plan 133: int → uint saturation (negative → 0, D54 precedent).
+ * Uses intptr_t/uintptr_t directly (nova_int/nova_uint are defined later in nova_rt.h).
  * int → u64 (direct cast) remains bit-cast; only `as uint` uses this. */
-static inline uint64_t nova_int_to_uint(int64_t x) {
-    return (x < 0) ? (uint64_t)0 : (uint64_t)x;
+static inline uintptr_t nova_int_to_uint(intptr_t x) {
+    return (x < 0) ? (uintptr_t)0 : (uintptr_t)x;
 }
 
 #endif /* NOVA_CAST_H */
